homework #6 excessive viewModel deleted, tabulations corrected, unused libraries removed
