+++ app/src/main/java/com/bignerdranch/android/androidacademy/data/MoviesLoad.kt
@@ -11,19 +11,19 @@ import retrofit2.http.Query
 
 class MoviesRep {
     suspend fun loadMovies(): Page {
-        return RetrofitModule.moviesApi.getMovies()
+        return RetrofitModule().moviesApi.getMovies()
     }
 
     suspend fun loadDetails(id: Long): MovieDetails {
-        return RetrofitModule.moviesApi.getMovieDetails(id)
+        return RetrofitModule().moviesApi.getMovieDetails(id)
     }
 
     suspend fun loadActors(id: Long): Credit {
-        return RetrofitModule.moviesApi.getActors(id)
+        return RetrofitModule().moviesApi.getActors(id)
     }
 }
 
-private interface MoviesApi {
+interface MoviesApi {
     @GET("movie/popular?language=en-US&page=1")
     suspend fun getMovies(@Query("api_key") key: String = apiKey): Page
 
@@ -38,18 +38,6 @@ private interface MoviesApi {
         @Path("movie_id") movieId: Long,
         @Query("api_key") key: String = apiKey
     ): Credit
-}
-
-private object RetrofitModule {
-    private val json = Json {
-        ignoreUnknownKeys = true
-    }
-    private val retrofit: Retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
-        .build()
-
-    val moviesApi: MoviesApi = retrofit.create()
 }
 
 const val apiKey = "71d3014534ead31bdf3e30983069d05f"
