+++ app/src/main/java/com/bignerdranch/android/androidacademy/data/RetrofitModule.kt
@@ -1,0 +1,19 @@
+package com.bignerdranch.android.androidacademy.data
+
+import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import kotlinx.serialization.json.Json
+import okhttp3.MediaType.Companion.toMediaType
+import retrofit2.Retrofit
+import retrofit2.create
+
+class RetrofitModule {
+    private val json = Json {
+        ignoreUnknownKeys = true
+    }
+    private val retrofit: Retrofit = Retrofit.Builder()
+        .baseUrl(BASE_URL)
+        .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
+        .build()
+
+    val moviesApi: MoviesApi = retrofit.create()
+}
