+++ app/src/main/res/layout-land/fragment_movies_list.xml
@@ -22,7 +22,7 @@
             android:textStyle="bold"
             app:layout_constraintTop_toTopOf="parent"
             android:drawablePadding="10dp"
-            android:layout_marginTop="40dp"
+            android:layout_marginTop="20dp"
             android:layout_marginStart="16dp"
             app:layout_constraintStart_toStartOf="parent"
             app:drawableStartCompat="@drawable/target" />
